backend/requirements.txt
Django>=5.0,<5.1
djangorestframework>=3.15,<4.0
django-cors-headers>=4.3,<5.0
django-grappelli>=3.0,<4.0

